BugBox/game_login/requirements.txt
streamlit
qrcode
Pillow
opencv-python
numpy
pyzbar
pandas
python-dotenv
bcrypt
